+++ ft_resolve_hex.c
@@ -6,13 +6,13 @@
 /*   By: idavoli- <idavoli-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/11 03:30:00 by idavoli-          #+#    #+#             */
-/*   Updated: 2021/10/13 02:17:02 by idavoli-         ###   ########.fr       */
+/*   Updated: 2021/10/14 20:44:30 by idavoli-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_resolve_hex(unsigned int n, int is_upper)
+int	ft_resolve_hex(unsigned long long n, int is_upper)
 {
 	int	i;
